requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Django==4.2.9
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
firebase-admin==6.4.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
httplib2==0.22.0
idna==3.6
msgpack==1.0.7
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pytz==2024.1
requests==2.31.0
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
